online_courses/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
sqlparse==0.5.3
pytz==2023.3  # если необходимо
gunicorn==20.1.0  # если планируешь деплой
django-cors-headers==3.13.0  # если нужно CORS
